+++ samples/SampleApp/ViewModels/UserSettingViewModel.cs
@@ -1,9 +1,14 @@
 ﻿using SampleApp.Services;
 
 namespace SampleApp.ViewModels;
-public partial class UserSettingViewModel(IThemeService themeService) : ObservableObject
+public partial class UserSettingViewModel : ObservableObject
 {
-    readonly IThemeService themeService = themeService;
+    readonly IThemeService themeService;
+    public UserSettingViewModel(IThemeService themeService)
+    {
+        this.themeService = themeService;
+        InitializeTheme();
+    }
 
     [ObservableProperty]
     bool isCheckedLight;
@@ -12,35 +17,21 @@ public partial class UserSettingViewModel(IThemeService themeService) : Observab
     bool isCheckedDark;
 
     [ObservableProperty]
-    bool isCheckedSystem = true;
-    partial void OnIsCheckedLightChanged(bool value)
-    {
-        var a = themeService.UserAppTheme;
+    bool isCheckedSystem;
 
-        if (value)
-        {
-            themeService.SetTheme(AppTheme.Light);
-        }
-        var b = themeService.UserAppTheme;
-    }
-    partial void OnIsCheckedDarkChanged(bool value)
-    {
-        var a = themeService.UserAppTheme;
+    partial void OnIsCheckedLightChanged(bool value) =>
+     themeService.SetTheme(value ? AppTheme.Light : themeService.UserAppTheme);
 
-        if (value)
-        {
-            themeService.SetTheme(AppTheme.Dark);
-        }
-        var b = themeService.UserAppTheme;
-    }
-    partial void OnIsCheckedSystemChanged(bool value)
-    {
-        var a = themeService.UserAppTheme;
+    partial void OnIsCheckedDarkChanged(bool value) =>
+        themeService.SetTheme(value ? AppTheme.Dark : themeService.UserAppTheme);
 
-        if (value)
-        {
-            themeService.SetTheme(AppTheme.Unspecified);
-        }
-        var b = themeService.UserAppTheme;
+    partial void OnIsCheckedSystemChanged(bool value) =>
+        themeService.SetTheme(value ? AppTheme.Unspecified : themeService.UserAppTheme);
+
+    void InitializeTheme()
+    {
+        IsCheckedLight = themeService.UserAppTheme == AppTheme.Light;
+        IsCheckedDark = themeService.UserAppTheme == AppTheme.Dark;
+        IsCheckedSystem = themeService.UserAppTheme == AppTheme.Unspecified;
     }
 }
