+++ samples/SampleApp/Views/AdvancedPage.xaml
@@ -6,8 +6,10 @@
     xmlns:controls="clr-namespace:SampleApp.Controls"
     xmlns:local="clr-namespace:SampleApp.ViewModels"
     xmlns:plugin="clr-namespace:Plugin.Maui.Calendar.Controls;assembly=Plugin.Maui.Calendar"
-    Unloaded="UnloadedHandler"
-    x:Name="advancedCalendarPage">
+    xmlns:toolkit="http://schemas.microsoft.com/dotnet/2022/maui/toolkit"
+    x:Name="advancedCalendarPage"
+    BackgroundColor="{toolkit:AppThemeResource BackgroundColor}"
+    Unloaded="UnloadedHandler">
     <ContentPage.Resources>
         <ResourceDictionary>
             <Style x:Key="daysLabelStyle" TargetType="Label">
