--- samples/SampleApp/Views/CalendarPickerPopup.xaml
@@ -1,57 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<rg:PopupPage    
-    x:Class="SampleApp.Views.CalendarPickerPopup"
-    xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
-    xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-    xmlns:rg="clr-namespace:Mopups.Pages;assembly=Mopups"
-    xmlns:plugin="clr-namespace:Plugin.Maui.Calendar.Controls;assembly=Plugin.Maui.Calendar"
-    xmlns:local="clr-namespace:SampleApp.ViewModels"
-    Unloaded="UnloadedHandler"
-    CloseWhenBackgroundIsClicked="False">
-    <rg:PopupPage.BindingContext>
-        <local:CalendarPickerPopupViewModel />
-    </rg:PopupPage.BindingContext>
-
-    <rg:PopupPage.Content>
-        <Border
-            BackgroundColor="White"
-            HorizontalOptions="Center"
-            VerticalOptions="Center">
-            <VerticalStackLayout
-                IsClippedToBounds="True"                
-                Spacing="3">
-                <Label
-                    FontAttributes="Bold"
-                    FontSize="Large"
-                    HorizontalOptions="Fill"
-                    HorizontalTextAlignment="Center"
-                    Text="Choose Date"
-                    VerticalOptions="Center" />
-                <plugin:Calendar
-                    x:Name="calendar"
-                    Padding="10,0"
-                    FooterSectionVisible="False"
-                    HorizontalOptions="Center"
-                    MaximumDate="{Binding MaximumDate}"
-                    MinimumDate="{Binding MinimumDate}"
-                    ShownDate="{Binding InitalDate}"
-                    ArrowsFontFamily="FontAwesomeSolid"
-                    SelectedDate="{Binding SelectedDate}"
-                    VerticalOptions="Center"
-                    WidthRequest="320" />
-                <FlexLayout
-                    JustifyContent="SpaceBetween">
-                    <Button
-                        Command="{Binding ClearCommand}"
-                        HorizontalOptions="End"
-                        Text="Clear" />
-                    <Button
-                        Command="{Binding SuccessCommand}"
-                        HorizontalOptions="End"
-                        Text="Select" />
-                    <Button Command="{Binding CancelCommand}" Text="Cancel" />
-                </FlexLayout>
-            </VerticalStackLayout>
-        </Border>
-    </rg:PopupPage.Content>
-</rg:PopupPage>
