+++ samples/SampleApp/Views/MultiSelectionPage.xaml
@@ -8,6 +8,7 @@
     xmlns:toolkit="http://schemas.microsoft.com/dotnet/2022/maui/toolkit"
     x:Name="multiCalendarPage"
     Title="MultiSelectionPage"
+    BackgroundColor="{toolkit:AppThemeResource BackgroundColor}"
     Unloaded="UnloadedHandler">
     <ContentPage.Resources>
         <ResourceDictionary>
