+++ samples/SampleApp/Views/PickerPopups/CalendarPickerPopup.xaml.cs
