+++ samples/SampleApp/Views/PickerPopups/CalendarRangePickerPopup.xaml.cs
