+++ samples/SampleApp/Views/PickerPopups/CalendarRangePickerPopupSelectedDates.xaml.cs
