+++ samples/SampleApp/Views/SimplePage.xaml
@@ -7,6 +7,7 @@
     xmlns:plugin="clr-namespace:Plugin.Maui.Calendar.Controls;assembly=Plugin.Maui.Calendar"
     xmlns:toolkit="http://schemas.microsoft.com/dotnet/2022/maui/toolkit"
     x:Name="simpleCalendarPage"
+    BackgroundColor="{toolkit:AppThemeResource BackgroundColor}"
     Unloaded="UnloadedHandler">
     <ContentPage.Resources>
         <ResourceDictionary>
