+++ samples/SampleApp/Views/WeekendCalendarPage.xaml
@@ -8,7 +8,8 @@
     xmlns:plugin="clr-namespace:Plugin.Maui.Calendar.Controls;assembly=Plugin.Maui.Calendar"
     xmlns:toolkit="http://schemas.microsoft.com/dotnet/2022/maui/toolkit"
     xmlns:vm="clr-namespace:SampleApp.ViewModels"
-    x:Name="weekendCalendarPage">
+    x:Name="weekendCalendarPage"
+    BackgroundColor="{toolkit:AppThemeResource BackgroundColor}">
     <ContentPage.Resources>
         <ResourceDictionary>
             <Style x:Key="daysLabelStyle" TargetType="Label">
